Fixed more than one tasks download for the same url. Fixed failed HTTP connection blocks the download thread. Fixed images wrong render when the width is not 2^n. Added imagesign_logo.png. Removed redundancy codes.

+++ src/main/java/top/nowandfuture/mod/imagesign/mixin/MixinSignTileEntity.java
@@ -25,7 +25,7 @@ public abstract class MixinSignTileEntity extends TileEntity implements IForgeTi
         super.remove();
         if(world != null && world.isRemote()){
             ImageLoadManager.INSTANCE.tryRemoveFromLoadingList(pos.toLong());
-            ImageFetcher.INSTANCE.removeByPos(pos);
+            ImageFetcher.INSTANCE.removeByPos(pos.toLong());
         }
     }
 
@@ -41,7 +41,7 @@ public abstract class MixinSignTileEntity extends TileEntity implements IForgeTi
         super.onChunkUnloaded();
         if(world != null && world.isRemote()){
             ImageLoadManager.INSTANCE.tryRemoveFromLoadingList(pos.toLong());
-            ImageFetcher.INSTANCE.removeByPos(pos);
+            ImageFetcher.INSTANCE.removeByPos(pos.toLong());
         }
     }
